refactor : 추가 정보 입력 모달 컴포넌트 분리
