+++ src/App.tsx
@@ -3,18 +3,13 @@ import PageRouter from "./routes/pageRouter"
 import UnAuthHeader from './components/Header/Header'
 
 function App() {
-  
-
   return (
       <div className="bg-background w-full min-h-screen flex justify-center items-center">
         <div className='relative bg-white w-[1200px] min-h-screen mx-auto'>
           <UnAuthHeader />
           <PageRouter />
         </div>
-      </div>
-    
-    
-    
+      </div>   
   )
 }
